leg-mqtt-simulator/requirements.txt
paho-mqtt>=2.0
schedule>=1.2
